Rename lib-->e3sm_to_cmip
Make main script an entry point
